requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy
pandas
imageio